--- client/src/features/interview-question/expected-question/index.ts
@@ -1,1 +1,0 @@
-export { default as ExpectedQuestionSelector } from './ExpectedQuestionSelector';
