--- client/src/features/interview-question/hooks/index.ts
@@ -1,1 +1,0 @@
-export { default as useExpectedPlayListLogic } from './useExpectedPlayListLogic';
