--- client/src/features/interview-question/index.ts
@@ -1,2 +1,0 @@
-export { default as UserQuestionPlayList } from './playlist/UserQuestionPlayList';
-export { default as ExpectedQuestionSelector } from './expected-question/ExpectedQuestionSelector';
