--- client/src/features/interview-question/playlist/index.ts
@@ -1,1 +1,0 @@
-export { default as UserQuestionPlayList } from './UserQuestionPlayList';
